requirements.txt
openfaas-sdk
requests
